Change MJFileModel to only collect terminal nodes which are not error nodes.
Error nodes have a negative start index, which would confuse our binary search. Also they are generally not that useful in our case, because we collect syntax errors which contain the error nodes anyway. The token search method is also changed to account for a failed parse and return an empty list in this case.