Fix errors when syntax errors exist.
When syntax errors are encountered during parsing, an error node is added to the parse tree. These nodes don't contain location information so we can't use them for highlighting, which crashes in that case. Now error nodes are filtered when searching for the left and right leaf so highlighting works in this case.